requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
feedparser==6.0.11
anthropic==0.40.0
python-dotenv==1.0.1
aiofiles==24.1.0
beautifulsoup4==4.12.3
lxml==5.3.0
yfinance==0.2.54
pdfplumber==0.11.4
python-multipart==0.0.12
